### 2020-11-19 #2846 @damienmarsic
Hi,

Suggested reviewers

AngryMaciek
xie186
BjornFJohansson
j-andrews7
aponsero
j-andrews7
sigven
AhmedArslan
Datakid
Tijeco

Best regards,

Damien

From: whedon
Sent: Thursday, November 19, 2020 01:26
To: openjournals/joss-reviews
Cc: Damien Marsic; Mention
Subject: [openjournals/joss-reviews] [PRE REVIEW]: Parent-map: analysis ofparental contributions to evolved or engineered protein or DNA sequences(#2846)

Submitting author: @damienmarsic (Damien Marsic)
Repository: https://github.com/damienmarsic/Parent-map
Version: 1.1.0
Editor: Pending
Reviewer: Pending
Managing EiC: Daniel S. Katz
⚠️ JOSS reduced service mode ⚠️
Due to the challenges of the COVID-19 pandemic, JOSS is currently operating in a "reduced service mode". You can read more about what that means in our blog post.
Author instructions
Thanks for submitting your paper to JOSS @damienmarsic. Currently, there isn't an JOSS editor assigned to your paper.
@damienmarsic if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, this list of people have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
Editor instructions
