### 2020-11-24 #2846 @csoneson
👋 @xie186, @BjornFJohansson, @j-andrews7 - would two of you be interested in reviewing this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues, following these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).

> Parent-map: analysis of parental contributions to evolved or engineered protein or DNA sequences
