### 2020-11-25 #2846 @csoneson
Just for completeness: @BjornFJohansson - I think we're all set on the reviewer side for this submission, but I hope we can get back to you for another submission in the future. Thanks for volunteering to review for JOSS!
