### 2020-11-25 #2864 @whedon
[REVIEW]: Parent-map: analysis of parental contributions to evolved or engineered protein or DNA sequences
**Submitting author:** @damienmarsic (<a href="http://orcid.org/0000-0003-0847-8095">Damien Marsic</a>)
**Repository:** <a href="https://github.com/damienmarsic/Parent-map" target ="_blank">https://github.com/damienmarsic/Parent-map</a>
**Version:** 1.1.2
