### 2020-11-25 #2864 @csoneson
@xie186, @j-andrews7 - This is the review thread for this submission. All of our communications will happen here from now on.

Please have a look at the "Reviewer instructions & questions" in the first comment above, which also contains your respective checklists. As you go over the submission, please check any items that you feel have been satisfied, and provide some comments or guidance for the author concerning the remaining ones. We aim for the review process to be fully completed within about 4-6 weeks but please make a start well ahead of this as JOSS reviews are by their nature iterative and any early feedback you may be able to provide to the author will be very helpful in meeting this schedule.

Feel free to ping me (@csoneson) if you have any questions. Thanks again for accepting to review this submission!
