### 2020-11-30 damienmarsic/Parent-map#1 @j-andrews7
Contribution guidelines, issue template, etc.
Hi @damienmarsic,

As part of your [JOSS review](https://github.com/openjournals/joss-reviews/issues/2864), I noticed that the repo currently lacks [community guidelines for contributing](https://docs.github.com/en/free-pro-team@latest/github/building-a-strong-community/setting-guidelines-for-repository-contributors), where users can find support or how to structure an issue, and a PR template.

I'd like to see at least the first of those added, though the others may be a good idea as well.
