### 2020-11-30 damienmarsic/Parent-map#1 @damienmarsic
Hi Jared,
Thanks for your feedback. I created a CONTRIBUTING.md file, as well as bug report and feature request templates, and a discussion list.
