### 2020-12-01 damienmarsic/Parent-map#2 @j-andrews7
Basic functionality tests
Basic functionality tests would be a large boon to this package and let me tick a box for your [review](https://github.com/openjournals/joss-reviews/issues/2864).

**Describe the solution you'd like**
Testing is easy to implement via [pytest](https://docs.pytest.org/en/stable/) and can be simple.

A handful of tests to show that stats are properly calculated, deletions/insertions are handled as expected, and alignment works are all that are needed.
