+++ parent-map.py
@@ -1,9 +1,7 @@
 #!/usr/bin/env python
-
-# parent-map version 1.1.0
-# Author: Damien Marsic, damien.marsic@aliyun.com
-# 2020-06-17
-# License: GNU General Public v3 (GPLv3)
+__version__='1.1.1'
+last_update='2020-12-02'
+author='Damien Marsic, damien.marsic@aliyun.com'
 
 import argparse
 from gooey import Gooey, GooeyParser
@@ -15,10 +13,12 @@ pd.set_option('display.max_rows', None)
 pd.set_option('display.max_columns', None)
 pd.set_option('display.width', None)
 cli=False
+minfrag=6
+seqtype=''
 if len(list(sys.argv))>1:
     cli=True
 if '-v' in sys.argv or '--version' in sys.argv:
-    print('\n  Parent-map version 1.1.0\n  Damien Marsic, 2020\n')
+    print('\n  Parent-map version '+__version__+'\n  '+author+'\n  Last updated: '+last_update+'\n')
     sys.exit()
 
 def parse_CLI():
@@ -45,10 +45,10 @@ def parse_CLI():
     'menuTitle': 'About Parent-map',
     'name': 'Parent-map',
     'description': 'Characterize protein sequence variants against possible parental sequences',
-    'version': '1.1.0',
+    'version': __version__,
     'copyright': '2020',
     'website': 'https://github.com/damienmarsic/Parent-map',
-    'developer': 'Damien Marsic',
+    'developer': author,
     'license': 'GNU General Public v3 (GPLv3)'}]},
     {'name':'Help','items':[{'type':'Link', 'menuTitle': 'Parent-map documentation','url': 'https://parent-map.readthedocs.io' }]}])
 def parse_GUI():
@@ -225,7 +225,7 @@ def refine(i,a,b,x,y,seq,ref,db,sd):
                     l+=j[1]-j[0]+1
     return db,sd
 
-def colorize(filename,cd,id):
+def colorize(filename,cd,id,display):
     if not id:
         print()
         temp=""
@@ -287,559 +287,324 @@ def colorize(filename,cd,id):
     g.close()
     f.close()
     print('  Colorized version of parental map file saved into file: '+html+'\n')
-    if args.DisplayResults:
+    if display:
         webbrowser.open_new_tab(html)
     return
 
-if not cli:
-    args=parse_GUI()
-else:
-    args=parse_CLI()
-fail=False
-top=''
-score=0
-results=[]
-cd={}
+def main():
+    if not cli:
+        args=parse_GUI()
+    else:
+        args=parse_CLI()
+    fail=False
+    top=''
+    score=0
+    results=[]
+    cd={}
+    seqtype=''
 
-# Check arguments
+    # Check arguments
 
-if args.Variants and not check_file(args.Variants,'Sequence',1):
-    sys.exit()
-if args.Parents and not check_file(args.Parents,'Parent',1):
-    sys.exit()
-if not check_int(args.NumSeq,1,'There should be at least one sequence to be analyzed!'):
-    sys.exit()
-if not check_int(args.MinFragLen,1,'Minimal fragment length should be >0!'):
-    sys.exit()
-if not check_int(args.MinOverlap,0,'Minimal overlap length should be >=0!'):
-    sys.exit()
-if not check_int(args.MaxNameSize,0,'Maximal size of sequence names can not be less than 0!'):
-    sys.exit()
-if not check_int(args.SeqChars,10,'There should be at least 10 sequence characters per line!'):
-    sys.exit()
-if not check_int(args.VRSides,0,'The number of characters each side of variable regions can not be less than 0!'):
-    sys.exit()
-if args.VRSides>args.SeqChars/3:
[…]
