### 2020-12-02 damienmarsic/Parent-map#2 @damienmarsic
Hi Jared,
I added a test_.py script which tests 4 functions (+ 2 additional functions called by one of them). I had to make changes to parent-map.py becasue it did not have a main function so pytest could not work. Before running the test_.py script you first need to update parent-map to the latest version 1.1.1 (already on bioconda).
