### 2020-12-02 damienmarsic/Parent-map#2 @damienmarsic
By the way, I was unaware of pytest and all this testing stuff. I thought the most important test was to use parent-map with real-life example. That's why I assembled a collection of  sequences from the relevant literature.
