### 2020-12-02 damienmarsic/Parent-map#2 @j-andrews7
> By the way, I was unaware of pytest and all this testing stuff. I thought the most important test was to use parent-map with real-life example. That's why I assembled a collection of sequences from the relevant literature.

Oh, the examples are critical, don't get me wrong! And you've done a good job with them. Automated testing is more for you and others who may want to contribute to development - tells you that functionally isn't being broken or altered when it shouldn't be. Imagine you want to change the code to add a feature, but it unknowingly alters the output in an unwanted manner. Now with these tests, you'd catch that and be able to address it. It's an important part of software dev and a good habit to have.

Thanks for adding them!
