### 2020-12-02 #2864 @j-andrews7
In general, this is a really nifty tool, and I expect it will get a solid amount of play for its simplicity, ease of use, and straight-forward output - really an ideal fit for JOSS. **parent-map** is well-documented - the input and output formats are easy to understand and thoroughly explained on the associated [readthedocs](https://parent-map.readthedocs.io/en/latest/index.html) site, as are all arguments. Installation is simple via `conda` or `pip` - there's even a basic Windows-only GUI.

The tool functions as described and is highly performant, completing its task nearly instantly. I quadrupled the input size and didn't notice a difference, so I presume it likely scales adequately as well. Multiple example inputs from several published studies are provided in the repo. The examples are well-documented and automated tests were added [during the review process](https://github.com/damienmarsic/Parent-map/issues/2). Output from the GUI and CLI are identical, as expected.

The software repo is in good shape and contribution guidelines and issue templates are present. A google support forum is also available.

The paper is generally well-written and easy to follow. **My last request** is to organize the sections in the paper. Right now, the only sections are the Summary and Statement of Need, so having Implementation and/or Usage sections in addition would be helpful. Otherwise, this is good to go from me.
