### 2020-12-17 #2864 @xie186
This is really a neat tool to analyze parental contributions to evolved or engineered protein or DNA sequences.

The documentation is very detailed. Example data is provided for easy testing.

Parent-map can be installed using either `conda` or `pip` which is fantastic given the fact the easy installation is important.
Only problem is that error below was given after successfully run 'pip install parent-map':
```
ModuleNotFoundError: No module named 'gooey'
```
My testing system is ubuntu 16.04.

GUI mode is a great advantage of `parent-map`. GUI mode makes more users to be able to use the tool.

I suggest you to use `github issue` to replace the `Discussion list` hosted using `google group`. I don't see any advantage of using 'google group' here.

The manuscript is well written. One suggestion to give an example to represent the accuracy of `parent-map`.
