+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2020, Damien Marsic'
 author = 'Damien Marsic'
 
 # The full version, including alpha/beta/rc tags
-release = '1.1.0'
+release = '1.1.2'
 
 
 # -- General configuration ---------------------------------------------------
