+++ README.md
@@ -10,6 +10,8 @@ Install from PyPI:
 ````
 pip install parent-map
 ````
+Note that you may need to individually install some dependencies. If you get an error message about missing modules when trying to run parent-map after installing it with pip, install them using pip install. Example: pip install gooey
+
 Install from Bioconda:
 ```
 conda install -c bioconda parent-map
@@ -22,8 +24,6 @@ Windows installer: https://sourceforge.net/projects/parent-map/
 
 Full documentation: https://parent-map.readthedocs.io/
 
-Discussion list: https://groups.google.com/g/parent-map
-
 [![install with bioconda](https://img.shields.io/badge/install%20with-bioconda-brightgreen.svg?style=flat)](http://bioconda.github.io/recipes/parent-map/README.html)
 
 [![Download parent-map](https://a.fsdn.com/con/app/sf-download-button)](https://sourceforge.net/projects/parent-map/files/latest/download)
@@ -31,5 +31,3 @@ Discussion list: https://groups.google.com/g/parent-map
 [![Parent-map documentation](https://img.shields.io/badge/Parent--map-Documentation-yellow)](https://parent-map.readthedocs.io/)
 
 <div itemscope itemtype="https://schema.org/Person"><a itemprop="sameAs" content="https://orcid.org/0000-0003-0847-8095" href="https://orcid.org/0000-0003-0847-8095" target="orcid.widget" rel="me noopener noreferrer" style="vertical-align:top;"><img src="https://orcid.org/sites/default/files/images/orcid_16x16.png" style="width:1em;margin-right:.5em;" alt="ORCID iD icon">https://orcid.org/0000-0003-0847-8095</a></div>
-
-
