### 2021-01-05 #2864 @damienmarsic
Hi everyone, and best wishes for the new year!
Sorry for the dealy, I was actally waiting for instructions on what to do next. Following the reviewers comments, I made the following changes in the repository:
- paper.md has now sections "Methods" and "Implementation" in addition to "Statement of need"
- The google group discussion list was removed rom README.md and replaced with a link to the repository's GitHub Issues
- Both README.md and the readthedocs documentations were updated with a warning to manually install possible missing modules such as gooey when installing with pip.
Sorry I am not sure to understand last comment by @xie186. In my mind, the accuracy is already demonstrated in the Examples section of the documentation.
Thanks all for your efforts!
