### 2021-01-05 #2864 @j-andrews7
This looks good to me and I have no further comments other than good job and thanks for the quick edits. Personally, I feel the accuracy is adequately addressed through the examples, and given the sparsity of other options and journal guidelines, that a true benchmarking study is beyond the scope of the journal.
