### 2021-01-14 #2864 @xie186
@damienmarsic Regarding the last comment, what I meant was to include the example(s) of the document in the manuscript. Otherwise the user won't have this information while read the MS.
