+++ paper.md
@@ -42,7 +42,7 @@ Parent-map was written under Python 3.7 as both a command-line interface (CLI) a
 
 Parent-map is a platform-independent Python script that generates a set of five output files from two input files. Input file names and options can be entered as arguments at launch time, resulting in parent-map running in CLI mode, or within the GUI, which starts if parent-map is launched without arguments. This flexibility allows parent-map to be deployed in a variety of settings, as a simple desktop application or even as a bioinformatics pipeline component. The first input file contains the variant sequences, typically the most frequent or the most enriched sequences obtained at the completion of a directed evolution experiment. The other input file is a set of potential parental sequences to the variant sequences. The most useful files generated by parent-map, particularly in the case of variants derived from DNA shuffling, are parental contribution maps (file names ending in –par.txt and –par.html, the latter being a colorized version of the former). Instead of all possible combinations, the simplest map that can accurately describe the variant is shown, using as few parents and as few fragments as possible. Other output files include a statistics file summarizing the variant sequences main features, a sequence definition file comprehensively defining each variant in terms of its parents, and an alignment file showing how variants differ from their common parent.
 
-Parent-map can be tested using the provided [variant](https://github.com/damienmarsic/Parent-map/blob/master/example_variants.fasta) and [parent](https://github.com/damienmarsic/Parent-map/blob/master/example_parents.fasta) sample files, based on available literature describing evolved and rationally designed AAV capsid variants. Variants AAV-DJ [@grimm_vitro_2008], AAV2.5T [@excoffon_directed_2009], NP84 [@paulk_bioengineered_2018] and OLIG001 [@powell_characterization_2016] are derived from shuffled DNA libraries. Variants AAV-F [@hanlon_selection_2019], AAV-PHP.B [@deverman_cre-dependent_2016], 7m8 [@dalkara_vivo-directed_2013] and rAAV2-retro [@tervo_designer_2016] are derived from peptide insertion libraries. Variants SCH2, SCH9 [@ojala_vivo_2018], LI-A and LI-C [@marsic_vector_2014] are derived from more complex rationally designed libraries. Variants AAV2i8 [@asokan_reengineering_2010] and AAV2-sept-Y-F [@petrs-silva_novel_2011] were rationally designed.
[…]
