### 2021-01-14 #2864 @damienmarsic
@xie186 Thanks for this clarification.
I added detailed information at the end of the manuscript, showing that parent-map correctly identifies all the features from the sample variants.
