### 2021-01-20 #2864 @csoneson
ping @xie186
> @xie186 - could you confirm whether your comments have been adequately addressed and you're happy with the current state of the submission (and if so, please tick the last boxes in the checklist above), or whether there are outstanding issues?
