### 2021-01-22 #2864 @csoneson
@damienmarsic - this looks all good. The next steps are as follows:

- [ ] Archive the software in Zenodo or a similar service
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (should match that of the paper). You may also add the author's ORCID.

I will then move forward with accepting the submission.
